Add version argument and move location
